Rev CNI dependencies to recent master.
Pick up API changes on way to v0.4 support.

Bit of refactoring to make it easier to support other platforms in future.
