+++ utils/utils.go
@@ -1,4 +1,4 @@
-// Copyright 2015 Tigera Inc
+// Copyright 2015-2018 Tigera Inc
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,14 +23,11 @@ import (
 	"os"
 	"regexp"
 	"strings"
-	"time"
 
 	"github.com/containernetworking/cni/pkg/skel"
 	cnitypes "github.com/containernetworking/cni/pkg/types"
 	"github.com/containernetworking/cni/pkg/types/current"
-	"github.com/containernetworking/plugins/pkg/ip"
 	"github.com/containernetworking/plugins/pkg/ipam"
-	"github.com/containernetworking/plugins/pkg/ns"
 	"github.com/projectcalico/cni-plugin/types"
 	"github.com/projectcalico/libcalico-go/lib/apiconfig"
 	api "github.com/projectcalico/libcalico-go/lib/apis/v3"
@@ -39,7 +36,6 @@ import (
 	cnet "github.com/projectcalico/libcalico-go/lib/net"
 	"github.com/projectcalico/libcalico-go/lib/options"
 	"github.com/sirupsen/logrus"
-	"github.com/vishvananda/netlink"
 )
 
 func Min(a, b int) int {
@@ -95,51 +91,6 @@ func CreateOrUpdate(ctx context.Context, client client.Interface, wep *api.Workl
 	}
 
 	return client.WorkloadEndpoints().Create(ctx, wep, options.SetOptions{})
-}
-
-// CleanUpNamespace deletes the devices in the network namespace.
-func CleanUpNamespace(args *skel.CmdArgs, logger *logrus.Entry) error {
-	// Only try to delete the device if a namespace was passed in.
-	if args.Netns != "" {
-		logger.WithFields(logrus.Fields{
-			"netns": args.Netns,
-			"iface": args.IfName,
-		}).Debug("Checking namespace & device exist.")
-		devErr := ns.WithNetNSPath(args.Netns, func(_ ns.NetNS) error {
-			_, err := netlink.LinkByName(args.IfName)
-			return err
-		})
-
-		if devErr == nil {
-			fmt.Fprintf(os.Stderr, "Calico CNI deleting device in netns %s\n", args.Netns)
-			// Deleting the veth has been seen to hang on some kernel version. Timeout the command if it takes too long.
-			ch := make(chan error, 1)
-
-			go func() {
-				err := ns.WithNetNSPath(args.Netns, func(_ ns.NetNS) error {
-					_, err := ip.DelLinkByNameAddr(args.IfName, netlink.FAMILY_V4)
-					return err
-				})
-
-				ch <- err
-			}()
-
-			select {
-			case err := <-ch:
-				if err != nil {
-					return err
-				} else {
-					fmt.Fprintf(os.Stderr, "Calico CNI deleted device in netns %s\n", args.Netns)
-				}
-			case <-time.After(5 * time.Second):
-				return fmt.Errorf("Calico CNI timed out deleting device in netns %s", args.Netns)
-			}
-		} else {
-			logger.WithField("ifName", args.IfName).Info("veth does not exist, no need to clean up.")
-		}
-	}
-
-	return nil
 }
 
 // CleanUpIPAM calls IPAM plugin to release the IP address.
